make helpers.go and tidy up code
